fix(astro): unblock SSR by lazy-loading echarts; mermaid + benchmark share dark hook
main 의 echarts-for-react CJS deep import 가 Vite SSR 평가에서 `exports is not defined` 로 500
유발. landing 과 reference/benchmarks 페이지 둘 다 영향.

대응:
- BenchmarkCharts.tsx 를 `lazy()` + Suspense wrapper 로 분리. 실제 echarts 코드는
  BenchmarkChartsImpl.tsx (rename) 에. 정적 import 가 사라지면서 SSR 모듈 평가 단계에서
  CJS body 가 evaluate 되지 않음. 부수 효과로 echarts chunk 가 페이지 진입 시점에만 fetch.
- Suspense fallback 의 minHeight 를 실제 panel 개수 × MIN_CHART_HEIGHT 합산으로 — 기존
  400px 으로는 차트 5개 그릴 자리 2100px+ 대비 CLS 1700px+ 발생.
- 새 hook `useStarlightDark` 로 BenchmarkChartsImpl 의 `useIsDark` 추출. Mermaid 도 동일
  hook 사용 — 그동안 mount 시점에 한 번만 themeVariables 가 박혀 다크 토글 무시하던 것을
  isDark 변화 시 재 initialize + render 하도록 수정.
- Mermaid theme tokens 를 [light, dark] paired 테이블로 묶어 한쪽 키 추가 누락 방지.
- 같은 페이지 N 개 다이어그램이 동시 mount 될 때 `mermaid.initialize` 가 N 번 중복 호출
  되던 부분을 `lastInitDark` module 가드로 N-1 번 skip.

검증:
- `bun run build` 통과, 513 페이지, starlight-links-validator 깨끗
- dev server 띄워 landing / reference/benchmarks / reference/napi (mermaid 포함) 모두 HTTP 200

skip (별도 PR 후보):
- mermaid/echarts/diagrams.ts 의 hardcoded brand hex 를 tailwind --color-* 와 일원화
  (현재 4 파일 중복)
- BenchmarkChartsImpl + MetafileAnalyzer 의 `echarts.use(...)` 통합
