+++ documents/src/components/Mermaid.tsx
@@ -1,5 +1,7 @@
 import { useEffect, useId, useRef, useState } from "react";
-import type { Mermaid as MermaidApi } from "mermaid";
+import type { Mermaid as MermaidApi, MermaidConfig } from "mermaid";
+
+import { useStarlightDark } from "./useStarlightDark";
 
 interface Props {
   /** mermaid diagram source (raw text — `flowchart TD`, `sequenceDiagram`, etc.) */
@@ -15,39 +17,57 @@ interface Props {
 let mermaidPromise: Promise<MermaidApi> | null = null;
 function loadMermaid(): Promise<MermaidApi> {
   if (mermaidPromise) return mermaidPromise;
-  mermaidPromise = import("mermaid").then(({ default: mermaid }) => {
-    mermaid.initialize({
-      startOnLoad: false,
-      theme: "base",
-      securityLevel: "loose",
-      fontFamily: "var(--sl-font, sans-serif)",
-      themeVariables: {
-        primaryColor: "#fff7ed",
-        primaryBorderColor: "#f7a41d",
-        primaryTextColor: "#431407",
-        secondaryColor: "#ffedd5",
-        secondaryBorderColor: "#fb923c",
-        tertiaryColor: "#fafaf9",
-        tertiaryBorderColor: "#d6d3d1",
-        clusterBkg: "#1c1816",
-        clusterBorder: "#3a3432",
-        titleColor: "#f5f5f4",
-        edgeLabelBackground: "#141110",
-        lineColor: "#f7a41d",
-        fontSize: "15px",
-      },
-      flowchart: {
-        htmlLabels: true,
-        curve: "basis",
-        nodeSpacing: 44,
-        rankSpacing: 58,
-        padding: 28,
-        useMaxWidth: true,
-      },
-    });
-    return mermaid;
-  });
+  mermaidPromise = import("mermaid").then(({ default: mermaid }) => mermaid);
   return mermaidPromise;
+}
+
+const COMMON_CONFIG: MermaidConfig = {
+  startOnLoad: false,
+  theme: "base",
+  securityLevel: "loose",
+  fontFamily: "var(--sl-font, sans-serif)",
+  flowchart: {
+    htmlLabels: true,
+    curve: "basis",
+    nodeSpacing: 44,
+    rankSpacing: 58,
+    padding: 28,
+    useMaxWidth: true,
+  },
+};
+
+// ZNTC 브랜드 오렌지 (#f7a41d) 는 양쪽 공통, 면/텍스트만 light↔dark 로 토글.
+// [light, dark] 페어로 묶어 한쪽 키만 추가하는 실수를 방지.
+const THEME_PAIRS = {
+  primaryColor: ["#fff7ed", "#1c1816"],
+  primaryBorderColor: ["#f7a41d", "#f7a41d"],
+  primaryTextColor: ["#431407", "#fed7aa"],
+  secondaryColor: ["#ffedd5", "#2a2422"],
+  secondaryBorderColor: ["#fb923c", "#fb923c"],
+  tertiaryColor: ["#fafaf9", "#141110"],
+  tertiaryBorderColor: ["#d6d3d1", "#3a3432"],
+  clusterBkg: ["#fafaf9", "#1c1816"],
+  clusterBorder: ["#d6d3d1", "#3a3432"],
+  titleColor: ["#1c1816", "#f5f5f4"],
+  edgeLabelBackground: ["#ffffff", "#141110"],
+  lineColor: ["#9a3412", "#f7a41d"],
+  fontSize: ["15px", "15px"],
+} as const;
+
+function themeVarsFor(isDark: boolean) {
+  const idx = isDark ? 1 : 0;
+  const out: Record<string, string> = {};
+  for (const [key, pair] of Object.entries(THEME_PAIRS)) out[key] = pair[idx];
+  return out;
+}
+
+// mermaid config 는 module-global. 같은 isDark 로 이미 init 됐으면 다이어그램 N 개의
+// 동시 mount 에서 N-1 번의 redundant initialize 를 skip — 결과는 동일하지만 work 절약.
+let lastInitDark: boolean | null = null;
+function ensureMermaidTheme(mermaid: MermaidApi, isDark: boolean): void {
+  if (lastInitDark === isDark) return;
+  mermaid.initialize({ ...COMMON_CONFIG, themeVariables: themeVarsFor(isDark) });
+  lastInitDark = isDark;
 }
 
 /** Client-side Mermaid renderer with lazy library load. `client:visible` 권장. */
@@ -56,11 +76,15 @@ export default function Mermaid({ chart }: Props) {
   const [error, setError] = useState<string | null>(null);
   const reactId = useId();
   const renderId = `mermaid-${reactId.replace(/:/g, "")}`;
+  const isDark = useStarlightDark();
 
   useEffect(() => {
     let cancelled = false;
     loadMermaid()
-      .then((mermaid) => mermaid.render(renderId, chart))
+      .then((mermaid) => {
+        ensureMermaidTheme(mermaid, isDark);
+        return mermaid.render(renderId, chart);
+      })
       .then(({ svg }) => {
         if (!cancelled && ref.current) ref.current.innerHTML = svg;
       })
@@ -70,7 +94,7 @@ export default function Mermaid({ chart }: Props) {
     return () => {
       cancelled = true;
     };
-  }, [chart, renderId]);
+  }, [chart, renderId, isDark]);
 
   if (error) {
     return (
